# lab4/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "BMSTU Open IT Platform. API для работы с услугами (dataGrowthFactor), заявками (growthRequest) и пользователями.",
        "title": "BITOP",
        "contact": {
            "name": "API Support",
            "url": "https://vk.com/bmstu_schedule",
            "email": "bitop@spatecon.ru"
        },
        "license": {
            "name": "AS IS (NO WARRANTY)"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/api/data-growth-factors": {
            "get": {
                "description": "Возвращает все услуги (факторы роста) с возможностью фильтрации по названию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data_growth_factors"
                ],
                "summary": "Получить список услуг",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.DataGrowthFactor"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт новую услугу (фактор роста)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data_growth_factors"
                ],
                "summary": "Добавить услугу",
                "parameters": [
                    {
                        "description": "Данные новой услуги",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.DataGrowthFactor"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ds.DataGrowthFactor"
                        }
                    }
                }
            }
        },
        "/api/data-growth-factors/{id}": {
            "get": {
                "description": "Возвращает услугу (фактор роста) по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data_growth_factors"
                ],
                "summary": "Получить услугу по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.DataGrowthFactor"
                        }
                    },
                    "404": {
                        "description": "Услуга не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет информацию об услуге (факторе роста)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data_growth_factors"
                ],
                "summary": "Обновить услугу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.DataGrowthFactor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет услугу (меняет флаг удаления или полностью удаляет запись)",
                "tags": [
                    "data_growth_factors"
                ],
                "summary": "Удалить услугу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DataGrowthFactor deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "data_growth_factor not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/data-growth-factors/{id}/add": {
            "post": {
                "description": "Добавляет выбранную услугу в текущую заявку пользователя со статусом \"черновик\"",
                "tags": [
                    "data_growth_factors"
                ],
                "summary": "Добавить услугу в заявку-черновик",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DataGrowthFactor added to draft successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/data-growth-factors/{id}/image": {
            "post": {
                "description": "Загружает изображение в MinIO и сохраняет URL в базе данных",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data_growth_factors"
                ],
                "summary": "Загрузить изображение услуги",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID услуги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "image_url",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/growth-request-data-growth-factors/{id}": {
            "put": {
                "description": "Обновляет числовое значение factor_num для DataGrowthFactor в корзине текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "growth_request_data_growth_factors"
                ],
                "summary": "Обновить значение factor_num",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фактора роста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет DataGrowthFactor из черновика текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "growth_request_data_growth_factors"
                ],
                "summary": "Удалить фактор роста из черновика",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фактора роста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.DataGrowthFactor": {
            "type": "object",
            "properties": {
                "attribute": {
                    "type": "string"
                },
                "coeff": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "isDelete": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}